+++ packages/logger/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add rolling file log handler
 
 ## 1.0.4 - 2021/9/9